+++ src/core/SessionInfo.h
@@ -79,11 +79,11 @@ public:
   explicit TSessionUI(TObjectClassId Kind) noexcept : TObject(Kind) {}
   virtual ~TSessionUI() noexcept = default;
   virtual void Information(const UnicodeString AStr, bool Status) = 0;
-  virtual uintptr_t QueryUser(const UnicodeString AQuery,
-    TStrings *MoreMessages, uintptr_t Answers, const TQueryParams *Params,
+  virtual uint32_t QueryUser(const UnicodeString AQuery,
+    TStrings *MoreMessages, uint32_t Answers, const TQueryParams *Params,
     TQueryType QueryType = qtConfirmation) = 0;
-  virtual uintptr_t QueryUserException(const UnicodeString AQuery,
-    Exception *E, uintptr_t Answers, const TQueryParams *Params,
+  virtual uint32_t QueryUserException(const UnicodeString AQuery,
+    Exception *E, uint32_t Answers, const TQueryParams *Params,
     TQueryType QueryType = qtConfirmation) = 0;
   virtual bool PromptUser(TSessionData *Data, TPromptKind Kind,
     const UnicodeString AName, const UnicodeString AInstructions, TStrings *Prompts,
