+++ src/windows/WinInterface.h
@@ -72,7 +72,7 @@ public:
   uintptr_t Timer{0};
   TQueryParamsTimerEvent TimerEvent;
   UnicodeString TimerMessage;
-  uintptr_t TimerAnswers{0};
+  uint32_t TimerAnswers{0};
   TQueryType TimerQueryType;
   uintptr_t Timeout{0};
   uintptr_t TimeoutAnswer{0};
@@ -129,21 +129,21 @@ void InitializeWinHelp();
 void FinalizeWinHelp();
 
 // windows\WinInterface.cpp
-uintptr_t MessageDialog(UnicodeString Msg, TQueryType Type,
-  uintptr_t Answers, UnicodeString HelpKeyword = HELP_NONE, const TMessageParams *Params = nullptr);
-uintptr_t MessageDialog(intptr_t Ident, TQueryType Type,
-  uintptr_t Answers, UnicodeString HelpKeyword = HELP_NONE, const TMessageParams *Params = nullptr);
-uintptr_t SimpleErrorDialog(UnicodeString Msg, UnicodeString MoreMessages = L"");
+uint32_t MessageDialog(UnicodeString Msg, TQueryType Type,
+  uint32_t Answers, UnicodeString HelpKeyword = HELP_NONE, const TMessageParams *Params = nullptr);
+uint32_t MessageDialog(intptr_t Ident, TQueryType Type,
+  uint32_t Answers, UnicodeString HelpKeyword = HELP_NONE, const TMessageParams *Params = nullptr);
+uint32_t SimpleErrorDialog(UnicodeString Msg, UnicodeString MoreMessages = L"");
 
-uintptr_t MoreMessageDialog(UnicodeString Message,
-  TStrings *MoreMessages, TQueryType Type, uintptr_t Answers,
+uint32_t MoreMessageDialog(UnicodeString Message,
+  TStrings *MoreMessages, TQueryType Type, uint32_t Answers,
   UnicodeString HelpKeyword, const TMessageParams *Params = nullptr);
 
-uintptr_t ExceptionMessageDialog(Exception *E, TQueryType Type,
-  UnicodeString MessageFormat = L"", uintptr_t Answers = qaOK,
+uint32_t ExceptionMessageDialog(Exception *E, TQueryType Type,
+  UnicodeString MessageFormat = L"", uint32_t Answers = qaOK,
   UnicodeString HelpKeyword = HELP_NONE, const TMessageParams *Params = nullptr);
-uintptr_t FatalExceptionMessageDialog(Exception *E, TQueryType Type,
-  intptr_t SessionReopenTimeout, UnicodeString MessageFormat = L"", uintptr_t Answers = qaOK,
+uint32_t FatalExceptionMessageDialog(Exception *E, TQueryType Type,
+  intptr_t SessionReopenTimeout, UnicodeString MessageFormat = L"", uint32_t Answers = qaOK,
   UnicodeString HelpKeyword = HELP_NONE, const TMessageParams *Params = nullptr);
 
 // forms\Custom.cpp
